Move receiveSyncPacket
