+++ src/client/java/net/pneumono/pneumonocore/config_api/ClientConfigApiRegistry.java
@@ -1,13 +1,18 @@
 package net.pneumono.pneumonocore.config_api;
 
+import com.mojang.datafixers.util.Pair;
+import com.mojang.serialization.DataResult;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
+import net.minecraft.nbt.NbtCompound;
+import net.minecraft.nbt.NbtElement;
+import net.minecraft.nbt.NbtOps;
 import net.pneumono.pneumonocore.PneumonoCore;
 import net.pneumono.pneumonocore.config_api.configurations.*;
 import net.pneumono.pneumonocore.config_api.screen.entries.*;
 
 public class ClientConfigApiRegistry {
     public static void register() {
-        ClientPlayNetworking.registerGlobalReceiver(ConfigSyncS2CPayload.ID, (payload, context) -> ClientConfigApi.receiveSyncPacket(payload));
+        ClientPlayNetworking.registerGlobalReceiver(ConfigSyncS2CPayload.ID, (payload, context) -> ClientConfigApiRegistry.receiveSyncPacket(payload));
 
         ClientConfigCommandRegistry.registerClientConfigCommand();
 
@@ -40,5 +45,33 @@ public class ClientConfigApiRegistry {
 
     private static void registerConfigEntryType(String name, ClientConfigApi.EntryFactory selector) {
         ClientConfigApi.registerConfigEntryType(PneumonoCore.identifier(name), selector);
+    }
+
+    public static void receiveSyncPacket(ConfigSyncS2CPayload payload) {
+        ConfigApi.LOGGER.info("Received config sync packet");
+
+        for (ConfigFile configFile : ConfigApi.getConfigFiles()) {
+            NbtCompound compound = payload.storedValues().getCompound(configFile.getModID()).orElse(null);
+            if (compound == null) continue;
+
+            for (AbstractConfiguration<?> configuration : configFile.getConfigurations()) {
+                NbtElement element = compound.get(configuration.info().getName());
+                if (element == null) continue;
+
+                if (!setReceivedEffectiveValue(configuration, element)) {
+                    ConfigApi.LOGGER.warn("Config sync packet contains invalid value '{}' for config '{}'. The default config value will be used instead.", element, configuration.info().getId());
+                }
+            }
+        }
+    }
+
+    private static  <T> boolean setReceivedEffectiveValue(AbstractConfiguration<T> config, NbtElement jsonElement) {
+        DataResult<Pair<T, NbtElement>> result = config.getValueCodec().decode(NbtOps.INSTANCE, jsonElement);
+        if (result.isError()) {
+            return false;
+        }
+
+        ConfigManager.setEffectiveValue(config, result.getOrThrow().getFirst());
+        return true;
     }
 }
